Enable back markdown rendering for the /@documentation link
